postメソッド切り分け
