+++ front/src/firestore/travelinks/index.ts
@@ -1,1 +1,2 @@
 export { useGetTravelinks } from '@/firestore/travelinks/useGetTravelinks'
+export { usePostTravelinks } from '@/firestore/travelinks/usePostTravelinks'
